Move credentials directory
