+++ pkg/credentials/azure/azure_secret.go
