+++ pkg/credentials/azure/azure_secret_test.go
