+++ pkg/credentials/s3/s3_secret_test.go
