+++ pkg/webhook/admission/pod/storage_initializer_injector_test.go
@@ -23,9 +23,9 @@ import (
 	"testing"
 
 	"github.com/kubeflow/kfserving/pkg/constants"
-	"github.com/kubeflow/kfserving/pkg/controller/inferenceservice/resources/credentials"
-	"github.com/kubeflow/kfserving/pkg/controller/inferenceservice/resources/credentials/gcs"
-	"github.com/kubeflow/kfserving/pkg/controller/inferenceservice/resources/credentials/s3"
+	"github.com/kubeflow/kfserving/pkg/credentials"
+	"github.com/kubeflow/kfserving/pkg/credentials/gcs"
+	"github.com/kubeflow/kfserving/pkg/credentials/s3"
 	"github.com/onsi/gomega"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
